requirements.txt
astronomer-cosmos==1.2.5
astro-run-dag==0.2.7
psycopg2-binary==2.9.9
apache-airflow-providers-airbyte==3.9.0

dbt-core==1.6.9
dbt-postgres==1.6.9
dbt-singlestore==1.6.0

dbt-snowflake>=1.5.0 



diff-match-patch==20200713
fastapi==0.88.0
python-json-logger==2.0.4
sse-starlette==0.9.0
gunicorn==20.1.0
uvicorn[standard]==0.18.3
websockets==10.0
psutil==5.9.2
requests==2.28.1
celery==5.2.7
redis==4.5.4